+++ cli/cli/commands/enclave/inspect/inspect.go
@@ -17,8 +17,9 @@ import (
 	"github.com/kurtosis-tech/kurtosis-cli/cli/helpers/logrus_log_levels"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/helpers/output_printers"
 	"github.com/kurtosis-tech/kurtosis-cli/commons/positional_arg_parser"
-	"github.com/kurtosis-tech/kurtosis-core/commons/enclave_object_labels"
-	"github.com/palantir/stacktrace"
+	"github.com/kurtosis-tech/object-attributes-schema-lib/forever_constants"
+	"github.com/kurtosis-tech/object-attributes-schema-lib/schema"
+	"github.com/kurtosis-tech/stacktrace"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 	"sort"
@@ -145,7 +146,7 @@ func run(cmd *cobra.Command, args []string) error {
 // ====================================================================================================
 func getEnclaveStatus(ctx context.Context, dockerManager *docker_manager.DockerManager, enclaveId string) (enclave_statuses.EnclaveStatus, error) {
 	searchLabels := map[string]string{
-		enclave_object_labels.EnclaveIDContainerLabel: enclaveId,
+		schema.EnclaveIDContainerLabel: enclaveId,
 	}
 	// TODO Replace with a call to the engine server!
 	enclaveContainers, err := dockerManager.GetContainersByLabels(ctx, searchLabels, shouldExamineStoppedContainersWhenPrintingEnclaveStatus)
@@ -168,9 +169,9 @@ func sortContainersByGUID(containers []*types.Container) ([]*types.Container, er
 	containersSet := map[string]*types.Container{}
 	for _, container := range containers {
 		if container != nil {
-			containerGUID, found := container.GetLabels()[enclave_object_labels.GUIDLabel]
+			containerGUID, found := container.GetLabels()[schema.GUIDLabel]
 			if !found {
-				return nil, stacktrace.NewError("No '%v' container label was found in container ID '%v' with labels '%+v'", enclave_object_labels.GUIDLabel, container.GetId(), container.GetLabels())
+				return nil, stacktrace.NewError("No '%v' container label was found in container ID '%v' with labels '%+v'", schema.GUIDLabel, container.GetId(), container.GetLabels())
 			}
 			containersSet[containerGUID] = container
 		}
@@ -182,7 +183,7 @@ func sortContainersByGUID(containers []*types.Container) ([]*types.Container, er
 	}
 
 	sort.Slice(containersResult, func(i, j int) bool {
-		return containersResult[i].GetLabels()[enclave_object_labels.GUIDLabel] < containersResult[j].GetLabels()[enclave_object_labels.GUIDLabel]
+		return containersResult[i].GetLabels()[schema.GUIDLabel] < containersResult[j].GetLabels()[schema.GUIDLabel]
 	})
 
 	return containersResult, nil
@@ -190,8 +191,8 @@ func sortContainersByGUID(containers []*types.Container) ([]*types.Container, er
 
 func getAPIContainerHostMachinePort(ctx context.Context, dockerManager *docker_manager.DockerManager, enclaveId string) (string, error) {
 	searchLabels := map[string]string{
-		enclave_object_labels.EnclaveIDContainerLabel: enclaveId,
-		enclave_object_labels.ContainerTypeLabel:      enclave_object_labels.ContainerTypeAPIContainer,
+		schema.EnclaveIDContainerLabel: enclaveId,
+		forever_constants.ContainerTypeLabel:      schema.ContainerTypeAPIContainer,
 	}
 	// TODO Replace with a call to the engine server!
 	enclaveContainers, err := dockerManager.GetContainersByLabels(ctx, searchLabels, shouldExamineStoppedContainersWhenPrintingEnclaveStatus)