+++ cli/cli/commands/engine/start/start.go
@@ -9,18 +9,21 @@ import (
 	"github.com/kurtosis-tech/kurtosis-cli/cli/defaults"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/helpers/engine_manager"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/helpers/logrus_log_levels"
-	"github.com/palantir/stacktrace"
+	"github.com/kurtosis-tech/object-attributes-schema-lib/schema"
+	"github.com/kurtosis-tech/stacktrace"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 	"strings"
 )
 
 const (
-	engineImageArg = "image"
-	logLevelArg    = "log-level"
+	engineVersionArg = "version"
+	logLevelArg      = "log-level"
+
+	defaultEngineVersion = ""
 )
 
-var engineImage string
+var engineVersion string
 var logLevelStr string
 
 var StartCmd = &cobra.Command{
@@ -32,10 +35,10 @@ var StartCmd = &cobra.Command{
 
 func init() {
 	StartCmd.Flags().StringVar(
-		&engineImage,
-		engineImageArg,
-		defaults.DefaultEngineImage,
-		"The image of the Kurtosis engine that should be started",
+		&engineVersion,
+		engineVersionArg,
+		defaultEngineVersion,
+		"The version (Docker tag) of the Kurtosis engine that should be started (blank will start the default version)",
 	)
 	StartCmd.Flags().StringVar(
 		&logLevelStr,
@@ -54,7 +57,7 @@ func init() {
 func run(cmd *cobra.Command, args []string) error {
 	ctx := context.Background()
 
-	logrus.Infof("Starting Kurtosis engine from image '%v'...", engineImage)
+	logrus.Infof("Starting Kurtosis engine from image '%v'...", engineVersion)
 
 	logLevel, err := logrus.ParseLevel(logLevelStr)
 	if err != nil {
@@ -71,11 +74,18 @@ func run(cmd *cobra.Command, args []string) error {
 	)
 
 	engineManager := engine_manager.NewEngineManager(dockerManager)
-	_, clientCloseFunc, err := engineManager.StartEngineIdempotently(ctx, engineImage, logLevel)
-	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred starting the Kurtosis engine")
+	objAttrsProvider := schema.GetObjectAttributesProvider()
+	var engineClientCloseFunc func() error
+	var startEngineErr error
+	if engineVersion == defaultEngineVersion {
+		_, engineClientCloseFunc, startEngineErr = engineManager.StartEngineIdempotentlyWithDefaultVersion(ctx, objAttrsProvider, logLevel)
+	} else {
+		_, engineClientCloseFunc, startEngineErr = engineManager.StartEngineIdempotentlyWithCustomVersion(ctx, objAttrsProvider, engineVersion, logLevel)
 	}
-	defer clientCloseFunc()
+	if startEngineErr != nil {
+		return stacktrace.Propagate(startEngineErr, "An error occurred starting the Kurtosis engine")
+	}
+	defer engineClientCloseFunc()
 
 	logrus.Info("Kurtosis engine started")