+++ cli/cli/defaults/defaults.go
@@ -7,8 +7,6 @@ package defaults
 
 import (
 	"github.com/kurtosis-tech/kurtosis-cli/cli/kurtosis_cli_version"
-	"github.com/kurtosis-tech/kurtosis-core/commons/kurtosis_core_version"
-	"github.com/kurtosis-tech/kurtosis-engine-server/engine/kurtosis_engine_server_version"
 	"github.com/sirupsen/logrus"
 )
 
@@ -17,8 +15,7 @@ const (
 
 	DefaultJavascriptReplImage = kurtosisDockerOrg + "/javascript-interactive-repl:" + kurtosis_cli_version.KurtosisCLIVersion
 
-	DefaultApiContainerImage = kurtosisDockerOrg + "/kurtosis-core_api:" + kurtosis_core_version.KurtosisCoreVersion
-
-	DefaultEngineImage = kurtosisDockerOrg + "/kurtosis-engine-server:" + kurtosis_engine_server_version.KurtosisEngineServerVersion
+	// If this version is passed to the engine, the engine will use its default version
+	DefaultAPIContainerVersion = ""
 )
 var DefaultEngineLogLevel = logrus.InfoLevel