+++ cli/cli/helpers/engine_manager/engine_existence_guarantor.go
@@ -2,28 +2,25 @@ package engine_manager
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"github.com/Masterminds/semver/v3"
 	"github.com/docker/go-connections/nat"
 	"github.com/kurtosis-tech/container-engine-lib/lib/docker_manager"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/command_str_consts"
-	"github.com/kurtosis-tech/kurtosis-cli/cli/helpers/engine_labels_schema"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/helpers/host_machine_directories"
-	"github.com/kurtosis-tech/kurtosis-engine-api-lib/golang/kurtosis_engine_api_version"
-	"github.com/kurtosis-tech/kurtosis-engine-api-lib/golang/kurtosis_engine_rpc_api_consts"
-	"github.com/kurtosis-tech/kurtosis-engine-server/engine/kurtosis_engine_server_docker_api"
-	"github.com/palantir/stacktrace"
+	"github.com/kurtosis-tech/kurtosis-engine-api-lib/api/golang/lib/kurtosis_context"
+	"github.com/kurtosis-tech/kurtosis-engine-server/launcher/engine_server_launcher"
+	"github.com/kurtosis-tech/object-attributes-schema-lib/schema"
+	"github.com/kurtosis-tech/stacktrace"
 	"github.com/sirupsen/logrus"
-	"time"
+	"os"
 )
 
 const (
-	containerNamePrefix = "kurtosis-engine"
+	// If set to empty, then we'll use whichever default version the launcher provides
+	defaultEngineImageVersionTag = ""
 
-	networkToStartEngineContainerIn = "bridge"
-
-	dockerSocketFilepath = "/var/run/docker.sock"
+	engineDataDirPermBits = 0755
 )
 
 // Visitor that does its best to guarantee that a Kurtosis engine is running
@@ -37,33 +34,61 @@ type engineExistenceGuarantor struct {
 
 	dockerManager *docker_manager.DockerManager
 
-	engineImage string
+	objAttrsProvider schema.ObjectAttributesProvider
 
-	engineAPIVersion string
+	engineServerLauncher *engine_server_launcher.EngineServerLauncher
+
+	imageVersionTag string
 
 	logLevel logrus.Level
+
+	// If an engine is currently running, then this will contain the version that the image is running with
+	// If no engine is running, this will be emptystring
+	maybeCurrentlyRunningEngineVersionTag string
 
 	// Port bindings of the engine server that is guaranteed to be started if the visiting didn't throw an error
 	// Will be nil before visiting
 	postVisitingHostMachinePortBinding *nat.PortBinding
 }
 
-func newEngineExistenceGuarantor(
+func newEngineExistenceGuarantorWithDefaultVersion(
 	ctx context.Context,
 	preVisitingMaybeHostMachinePortBinding *nat.PortBinding,
 	dockerManager *docker_manager.DockerManager,
-	engineImage string,
-	engineAPIVersion string,
+	objAttrsProvider schema.ObjectAttributesProvider,
 	logLevel logrus.Level,
+	maybeCurrentlyRunningEngineVersionTag string,
+) *engineExistenceGuarantor {
+	return newEngineExistenceGuarantorWithCustomVersion(
+		ctx,
+		preVisitingMaybeHostMachinePortBinding,
+		dockerManager,
+		objAttrsProvider,
+		defaultEngineImageVersionTag,
+		logLevel,
+		maybeCurrentlyRunningEngineVersionTag,
+	)
+}
+
+func newEngineExistenceGuarantorWithCustomVersion(
+	ctx context.Context,
+	preVisitingMaybeHostMachinePortBinding *nat.PortBinding,
+	dockerManager *docker_manager.DockerManager,
+	objAttrsProvider schema.ObjectAttributesProvider,
+	imageVersionTag string,
+	logLevel logrus.Level,
+	maybeCurrentlyRunningEngineVersionTag string,
 ) *engineExistenceGuarantor {
 	return &engineExistenceGuarantor{
 		ctx:                                    ctx,
 		preVisitingMaybeHostMachinePortBinding: preVisitingMaybeHostMachinePortBinding,
 		dockerManager:                          dockerManager,
-		engineImage:                            engineImage,
-		engineAPIVersion: engineAPIVersion,
+		objAttrsProvider:                       objAttrsProvider,
+		engineServerLauncher:                   engine_server_launcher.NewEngineServerLauncher(dockerManager, objAttrsProvider),
+		imageVersionTag:                        imageVersionTag,
 		logLevel:                               logLevel,
-		postVisitingHostMachinePortBinding:     nil,
+		maybeCurrentlyRunningEngineVersionTag:  maybeCurrentlyRunningEngineVersionTag,
+		postVisitingHostMachinePortBinding:     nil,  // Will be filled in upon visiting
 	}
 }
 
@@ -73,90 +98,40 @@ func (guarantor *engineExistenceGuarantor) getPostVisitingHostMachinePortBinding
 
 // If the engine is stopped, try to start it
 func (guarantor *engineExistenceGuarantor) VisitStopped() error {
-	logrus.Infof(
-		"No Kurtosis engine was found; attempting to start one using image '%v'...",
-		guarantor.engineImage,
-	)
-	matchingNetworks, err := guarantor.dockerManager.GetNetworksByName(guarantor.ctx, networkToStartEngineContainerIn)
-	if err != nil {
-		return stacktrace.Propagate(
-			err,
-			"An error occurred getting networks matching the network we want to start the engine in, '%v'",
-			networkToStartEngineContainerIn,
-		)
-	}
-	numMatchingNetworks := len(matchingNetworks)
-	if numMatchingNetworks == 0 && numMatchingNetworks > 1 {
-		return stacktrace.NewError(
-			"Expected exactly one network matching the name of the network that we want to start the engine in, '%v', but got %v",
-			networkToStartEngineContainerIn,
-			numMatchingNetworks,
-		)
-	}
-	targetNetwork := matchingNetworks[0]
-	targetNetworkId := targetNetwork.GetId()
-
-	containerStartTimeUnixSecs := time.Now().Unix()
-	containerName := fmt.Sprintf(
-		"%v_%v",
-		containerNamePrefix,
-		containerStartTimeUnixSecs,
-	)
-	enginePortObj, err := nat.NewPort(
-		kurtosis_engine_rpc_api_consts.ListenProtocol,
-		fmt.Sprintf("%v", kurtosis_engine_rpc_api_consts.ListenPort),
-	)
-	if err != nil {
-		return stacktrace.Propagate(
-			err,
-			"An error occurred creating a port object with port num '%v' and protocol '%v' to represent the engine's port",
-			kurtosis_engine_rpc_api_consts.ListenPort,
-			kurtosis_engine_rpc_api_consts.ListenProtocol,
-		)
-	}
-
+	logrus.Infof("No Kurtosis engine was found; attempting to start one...")
 	engineDataDirpath, err := host_machine_directories.GetEngineDataDirpath()
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred getting the engine data dirpath")
 	}
 
-	envVars, err := getEngineEnvVars(guarantor.logLevel, engineDataDirpath)
-	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred getting the engine envvars")
+	// NOTE: We create this in advance because if we leave
+	if err := os.MkdirAll(engineDataDirpath, engineDataDirPermBits); err != nil {
+		return stacktrace.Propagate(err, "An error occurred creating the engine data dirpath '%v'", engineDataDirpath)
 	}
 
-	usedPorts := map[nat.Port]docker_manager.PortPublishSpec{
-		enginePortObj: docker_manager.NewManualPublishingSpec(kurtosis_engine_rpc_api_consts.ListenPort),
+	var hostMachinePortBinding *nat.PortBinding
+	var engineLaunchErr error
+	if guarantor.imageVersionTag == defaultEngineImageVersionTag {
+		hostMachinePortBinding, engineLaunchErr = guarantor.engineServerLauncher.LaunchWithDefaultVersion(
+			guarantor.ctx,
+			guarantor.logLevel,
+			kurtosis_context.DefaultKurtosisEngineServerPortNum,
+			engineDataDirpath,
+		)
+	} else {
+		hostMachinePortBinding, engineLaunchErr = guarantor.engineServerLauncher.LaunchWithCustomVersion(
+			guarantor.ctx,
+			guarantor.imageVersionTag,
+			guarantor.logLevel,
+			kurtosis_context.DefaultKurtosisEngineServerPortNum,
+			engineDataDirpath,
+		)
+	}
+	if engineLaunchErr != nil {
+		return stacktrace.Propagate(engineLaunchErr, "An error occurred launching the engine server container")
 	}
 
-	bindMounts := map[string]string{
-		// Necessary so that the engine server can interact with the Docker engine
-		dockerSocketFilepath: dockerSocketFilepath,
-		engineDataDirpath: kurtosis_engine_server_docker_api.EngineDataDirpathOnEngineContainer,
-	}
-	createAndStartArgs := docker_manager.NewCreateAndStartContainerArgsBuilder(
-		guarantor.engineImage,
-		containerName,
-		targetNetworkId,
-	).WithEnvironmentVariables(
-		envVars,
-	).WithBindMounts(
-		bindMounts,
-	).WithUsedPorts(
-		usedPorts,
-	).WithLabels(
-		engine_labels_schema.EngineContainerLabels,
-	).Build()
-
-	_, hostMachinePortBindings, err := guarantor.dockerManager.CreateAndStartContainer(guarantor.ctx, createAndStartArgs)
-	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred starting the Kurtosis engine container")
-	}
-	hostMachineEnginePortBinding, found := hostMachinePortBindings[enginePortObj]
-	if !found {
-		return stacktrace.NewError("The Kurtosis engine server started successfully, but no host machine port binding was found")
-	}
-	guarantor.postVisitingHostMachinePortBinding = hostMachineEnginePortBinding
+	guarantor.postVisitingHostMachinePortBinding = hostMachinePortBinding
 	logrus.Info("Successfully started Kurtosis engine")
 	return nil
 }
@@ -190,49 +165,42 @@ func (guarantor *engineExistenceGuarantor) VisitRunning() error {
 // ====================================================================================================
 //                                      Private Helper Functions
 // ====================================================================================================
-func getEngineEnvVars(logLevel logrus.Level, engineDataDirpathOnHostMachine string) (map[string]string, error) {
-	// TODO replace with a constructor
-	args := kurtosis_engine_server_docker_api.EngineServerArgs{
-		LogLevelStr:                    logLevel.String(),
-		EngineDataDirpathOnHostMachine: engineDataDirpathOnHostMachine,
-	}
-	serializedBytes, err := json.Marshal(args)
-	if err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred JSON-serializing engine args '%+v'", args)
-	}
-	result := map[string]string{
-		kurtosis_engine_server_docker_api.SerializedArgsEnvVar: string(serializedBytes),
-	}
-	return result, nil
-}
-
 func (guarantor *engineExistenceGuarantor) checkIfEngineIsUpToDate() {
-
 	runningEngineSemver, cliEngineSemver, err := guarantor.getRunningAndCLIEngineVersions()
 	if err != nil {
-		logrus.Warn("An error occurred verifying that the running engine is on the latest version; you may be running an out-of-date engine version")
+		logrus.Warn("An error occurred getting the running engine's version; you may be running an out-of-date engine version")
 		logrus.Debugf("Getting running and CLI engine versions error: %v", err)
+		return
 	}
 
 	if runningEngineSemver.LessThan(cliEngineSemver) {
 		kurtosisRestartCmd := fmt.Sprintf("%v %v %v", command_str_consts.KurtosisCmdStr, command_str_consts.EngineCmdStr, command_str_consts.EngineRestartCmdStr)
-		logrus.Warningf("The currently-running Kurtosis engine version is '%v', but the latest version is '%v'; to restart the engine with the latest version use '%v'", guarantor.engineAPIVersion, kurtosis_engine_api_version.KurtosisEngineApiVersion, kurtosisRestartCmd)
+		logrus.Warningf(
+			"The currently-running Kurtosis engine version is '%v', but the latest version is '%v'; to restart the engine with the latest version use '%v'",
+			runningEngineSemver.String(),
+			cliEngineSemver.String(),
+			kurtosisRestartCmd,
+		)
 	} else {
-		logrus.Debugf("Currently running engine version '%v' which is up-to-date", guarantor.engineAPIVersion)
+		logrus.Debugf("Currently running engine version '%v' which is up-to-date", guarantor.maybeCurrentlyRunningEngineVersionTag)
 	}
 	return
 }
 
 func (guarantor *engineExistenceGuarantor) getRunningAndCLIEngineVersions() (*semver.Version, *semver.Version, error) {
-	runningEngineSemver, err := semver.StrictNewVersion(guarantor.engineAPIVersion)
+	if guarantor.maybeCurrentlyRunningEngineVersionTag == "" {
+		return nil, nil, stacktrace.NewError("Needed to report the currently-running engine's version, but it's emptystring")
+	}
+	runningEngineSemver, err := semver.StrictNewVersion(guarantor.maybeCurrentlyRunningEngineVersionTag)
 	if err != nil {
-		return nil, nil, stacktrace.Propagate(err, "An error occurred parsing running engine version string '%v' to semantic version", guarantor.engineAPIVersion)
+		return nil, nil, stacktrace.Propagate(err, "An error occurred parsing running engine version string '%v' to semantic version", guarantor.maybeCurrentlyRunningEngineVersionTag)
 	}
 
-	kurtosisEngineAPISemver, err := semver.StrictNewVersion(kurtosis_engine_api_version.KurtosisEngineApiVersion)
+	launcherEngineSemverStr := engine_server_launcher.DefaultVersion
+	launcherEngineSemver, err := semver.StrictNewVersion(launcherEngineSemverStr)
 	if err != nil {
-		return nil, nil, stacktrace.Propagate(err, "An error occurred parsing own engine version string '%v' to semantic version", kurtosis_engine_api_version.KurtosisEngineApiVersion)
+		return nil, nil, stacktrace.Propagate(err, "An error occurred parsing CLI's engine version string '%v' to semantic version", launcherEngineSemverStr)
 	}
 
-	return runningEngineSemver, kurtosisEngineAPISemver, nil
+	return runningEngineSemver, launcherEngineSemver, nil
 }